+++ components/dashboard/DataCard.tsx
@@ -1,3 +1,4 @@
+import { useMemo } from 'react'
 import { Card, CardContent, Skeleton, Typography } from '@mui/material'
 
 import styles from '../../styles/Home.module.css'
@@ -13,71 +14,74 @@ type Props = {
 
 const DataCard = ({ data, type, title }: Props): JSX.Element => {
   const { showExpiredDelegates } = useDashboard()
-  const cardData = data && [
-    {
-      name: 'Recognized',
-      delegates: data.filter((delegate) => delegate.status === 'recognized'),
-    },
-    {
-      name: 'Shadow',
-      delegates: data.filter((delegate) =>
-        showExpiredDelegates
-          ? delegate.status !== 'recognized'
-          : delegate.status === 'shadow'
-      ),
-    },
-    {
-      name: 'Total',
-      delegates: data.filter((delegate) =>
-        showExpiredDelegates ? true : delegate.status !== 'expired'
-      ),
-    },
-  ]
+  return useMemo(() => {
+    const cardData = data && [
+      {
+        name: 'Recognized',
+        delegates: data.filter((delegate) => delegate.status === 'recognized'),
+      },
+      {
+        name: 'Shadow',
+        delegates: data.filter((delegate) =>
+          showExpiredDelegates
+            ? delegate.status !== 'recognized'
+            : delegate.status === 'shadow'
+        ),
+      },
+      {
+        name: 'Total',
+        delegates: data.filter((delegate) =>
+          showExpiredDelegates ? true : delegate.status !== 'expired'
+        ),
+      },
+    ]
 
-  return (
-    <div className={styles.infoCard}>
-      <Typography
-        component='h3'
-        variant='h6'
-        gutterBottom
-        sx={{ color: (theme) => theme.palette.text.primary }}
-      >
-        {title}
-      </Typography>
-      <Card sx={{ height: 'calc(100% - 39px)' }}>
-        <CardContent style={{ padding: '0.8em', height: '100%' }}>
-          <div className={styles.infoCardContainer}>
-            {!cardData
-              ? Array.from(Array(3).keys()).map((key) => (
-                  <div
-                    key={key}
-                    className={styles.thirdWidth}
-                    style={{ padding: '0 0.2em' }}
-                  >
-                    <Skeleton animation='wave' height={60} />
-                  </div>
-                ))
-              : cardData.map(({ name, delegates }, index) => (
-                  <div key={index} className={styles.thirdWidth}>
-                    <p className={styles.infoCardValue}>
-                      {type === 'delegates-count'
-                        ? delegates.reduce(
-                            (acum, current) =>
-                              acum.add(current.name || current.voteDelegate),
-                            new Set()
-                          ).size
-                        : type === 'delegators-count'
-                        ? reduceDelegators(delegates)
-                        : reduceAndFormatDelegations(delegates)}
-                    </p>
-                    <p className={styles.infoCardLabel}>{name}</p>
-                  </div>
-                ))}
-          </div>
-        </CardContent>
-      </Card>
-    </div>
-  )
+    return (
+      <div className={styles.infoCard}>
+        <Typography
+          component='h3'
+          variant='h6'
+          gutterBottom
+          sx={{ color: (theme) => theme.palette.text.primary }}
+        >
+          {title}
+        </Typography>
+        <Card sx={{ height: 'calc(100% - 39px)' }}>
+          <CardContent style={{ padding: '0.8em', height: '100%' }}>
+            <div className={styles.infoCardContainer}>
+              {!cardData
+                ? Array.from(Array(3).keys()).map((key) => (
+                    <div
+                      key={key}
+                      className={styles.thirdWidth}
+                      style={{ padding: '0 0.2em' }}
+                    >
+                      <Skeleton animation='wave' height={60} />
+                    </div>
+                  ))
+                : cardData.map(({ name, delegates }, index) => (
+                    <div key={index} className={styles.thirdWidth}>
+                      <p className={styles.infoCardValue}>
+                        {type === 'delegates-count'
+                          ? delegates.reduce(
+                              (acum, current) =>
+                                acum.add(current.name || current.voteDelegate),
+                              new Set()
+                            ).size
+                          : type === 'delegators-count'
+                          ? reduceDelegators(delegates)
+                          : reduceAndFormatDelegations(delegates)}
+                      </p>
+                      <p className={styles.infoCardLabel}>{name}</p>
+                    </div>
+                  ))}
+            </div>
+          </CardContent>
+        </Card>
+      </div>
+    )
+    // eslint-disable-next-line
+  }, [showExpiredDelegates, data])
 }
 
 export default DataCard
