+++ components/dashboard/LineCharts/DelegateWeightsLineChart.tsx
@@ -1,3 +1,4 @@
+import { useMemo } from 'react'
 import { useDashboard } from '../../../context/DashboardContext'
 import { DelegateBalance, DelegateBalances } from '../../../lib/types/delegate'
 import LineChart from './LineChart'
@@ -13,31 +14,34 @@ const DelegateWeightsLineChart = ({
 }: Props) => {
   const { setSelectedTime } = useDashboard()
 
-  return (
-    <LineChart
-      data={
-        recognizedDelegates &&
-        delegatesBalancesData &&
-        recognizedDelegates.map((del) => ({
-          id: del.voteDelegate,
-          data: delegatesBalancesData.map((entry) => ({
-            x: entry.time,
-            y:
-              entry.balances.find((bal) => bal.address === del.voteDelegate)
-                ?.amount || 0,
-          })),
-        }))
-      }
-      legendX='Date'
-      legendY='MKR'
-      title='Recognized Delegates Vote Weights'
-      mkrColors={false}
-      enableClick={true}
-      clickFunction={setSelectedTime}
-      enableLegend={false}
-      enableArea={true}
-      stacked={true}
-    />
+  return useMemo(
+    () => (
+      <LineChart
+        data={
+          recognizedDelegates &&
+          delegatesBalancesData &&
+          recognizedDelegates.map((del) => ({
+            id: del.voteDelegate,
+            data: delegatesBalancesData.map((entry) => ({
+              x: entry.time,
+              y:
+                entry.balances.find((bal) => bal.address === del.voteDelegate)
+                  ?.amount || 0,
+            })),
+          }))
+        }
+        legendX='Date'
+        legendY='MKR'
+        title='Recognized Delegates Vote Weights'
+        mkrColors={false}
+        enableClick={true}
+        clickFunction={setSelectedTime}
+        enableLegend={false}
+        enableArea={true}
+        stacked={true}
+      />
+    ),
+    [delegatesBalancesData, recognizedDelegates, setSelectedTime]
   )
 }
