+++ components/dashboard/LineCharts/LineChart.tsx
@@ -1,8 +1,8 @@
-import { Dispatch, SetStateAction } from 'react'
+import { Dispatch, SetStateAction, useMemo } from 'react'
 import { Serie, ResponsiveLine } from '@nivo/line'
 import { Card, Skeleton, useTheme, Typography } from '@mui/material'
 
-import InfoTooltip from '../../InfoTooltip'
+import InfoTooltip from '../InfoTooltip'
 import styles from '../../../styles/Home.module.css'
 import { kFormatter, kFormatterInt } from '../../../lib/helpers'
 import getTheme from '../../../lib/nivo/theme'
@@ -43,105 +43,109 @@ const LineChart = ({
   const { selectedStartDate, selectedEndDate } = useDashboard()
   const theme = useTheme()
 
-  const chartData = data?.map((serie) => ({
-    ...serie,
-    data: serie.data.filter(
-      ({ x }) =>
-        x &&
-        (selectedStartDate ? x >= new Date(selectedStartDate) : true) &&
-        (selectedEndDate ? x <= new Date(selectedEndDate) : true)
-    ),
-  }))
+  return useMemo(() => {
+    const chartData = data?.map((serie) => ({
+      ...serie,
+      data: serie.data.filter(
+        ({ x }) =>
+          x &&
+          (selectedStartDate ? x >= new Date(selectedStartDate) : true) &&
+          (selectedEndDate ? x <= new Date(selectedEndDate) : true)
+      ),
+    }))
 
-  return (
-    <div className={styles.chartCard}>
-      <Typography
-        component='h3'
-        variant='h6'
-        gutterBottom
-        sx={{ color: (theme) => theme.palette.text.primary }}
-      >
-        {title} {infoTooltipText ? <InfoTooltip text={infoTooltipText} /> : ''}
-      </Typography>
-      <Card className={styles.chartContainer}>
-        {!chartData ? (
-          <Skeleton variant='rectangular' height={'100%'} animation='wave' />
-        ) : (
-          <ResponsiveLine
-            data={chartData}
-            xScale={{
-              type: 'time',
-              format: '%Y-%m-%dT%H:%M:%SZ',
-            }}
-            yScale={{ type: 'linear', stacked }}
-            xFormat='time:%b %d, %Y'
-            yFormat={(value) => kFormatter(+value, 2)}
-            margin={{
-              left: margin?.left || 60,
-              bottom: margin?.bottom || 45,
-              top: margin?.top || 5,
-              right: margin?.right || 80,
-            }}
-            theme={getTheme(theme)}
-            colors={mkrColors ? mkrPaletteMain : mkrPalette}
-            enablePoints={false}
-            enableGridX={false}
-            axisLeft={{
-              legend: legendY,
-              legendOffset: -50,
-              legendPosition: 'middle',
-              format: (value) => kFormatterInt(value),
-            }}
-            axisBottom={{
-              legend: legendX,
-              legendOffset: 36,
-              legendPosition: 'middle',
-              tickValues: 'every 3 months',
-              format: '%b %d, %Y',
-            }}
-            isInteractive={true}
-            useMesh={true}
-            tooltip={({ point }) => {
-              return (
-                <Card className={styles.chartTooltip}>
-                  <span
-                    className={styles.tooltipCircle}
-                    style={{ backgroundColor: point.color }}
-                  ></span>
-                  <span>
-                    {point.data.xFormatted}: <b>{point.data.yFormatted}</b>
-                  </span>
-                </Card>
-              )
-            }}
-            legends={
-              chartData.length === 1 || !enableLegend
-                ? []
-                : [
-                    {
-                      anchor: 'right',
-                      direction: 'column',
-                      itemWidth: 80,
-                      itemHeight: 20,
-                      translateX: 90,
-                      symbolSize: 10,
-                      symbolShape: 'circle',
-                    },
-                  ]
-            }
-            enableArea={enableArea}
-            areaOpacity={0.05}
-            enableSlices={enableSlices && 'x'}
-            onClick={(point) => {
-              if (!enableClick) return
-              const newVal = new Date(point.data.x).getTime()
-              if (clickFunction) clickFunction(newVal)
-            }}
-          />
-        )}
-      </Card>
-    </div>
-  )
+    return (
+      <div className={styles.chartCard}>
+        <Typography
+          component='h3'
+          variant='h6'
+          gutterBottom
+          sx={{ color: (theme) => theme.palette.text.primary }}
+        >
+          {title}
+          {infoTooltipText ? <InfoTooltip text={infoTooltipText} /> : ''}
+        </Typography>
+        <Card className={styles.chartContainer}>
+          {!chartData ? (
+            <Skeleton variant='rectangular' height={'100%'} animation='wave' />
+          ) : (
+            <ResponsiveLine
+              data={chartData}
+              xScale={{
+                type: 'time',
+                format: '%Y-%m-%dT%H:%M:%SZ',
+              }}
+              yScale={{ type: 'linear', stacked }}
+              xFormat='time:%b %d, %Y'
+              yFormat={(value) => kFormatter(+value, 2)}
+              margin={{
+                left: margin?.left || 60,
+                bottom: margin?.bottom || 45,
+                top: margin?.top || 5,
+                right: margin?.right || 80,
+              }}
+              theme={getTheme(theme)}
+              colors={mkrColors ? mkrPaletteMain : mkrPalette}
+              enablePoints={false}
+              enableGridX={false}
+              axisLeft={{
+                legend: legendY,
+                legendOffset: -50,
+                legendPosition: 'middle',
+                format: (value) => kFormatterInt(value),
+              }}
+              axisBottom={{
+                legend: legendX,
+                legendOffset: 36,
+                legendPosition: 'middle',
+                tickValues: 'every 3 months',
+                format: '%b %d, %Y',
+              }}
+              isInteractive={true}
+              useMesh={true}
+              tooltip={({ point }) => {
+                return (
+                  <Card className={styles.chartTooltip}>
+                    <span
+                      className={styles.tooltipCircle}
+                      style={{ backgroundColor: point.color }}
+                    ></span>
+                    <span>
+                      {point.data.xFormatted}: <b>{point.data.yFormatted}</b>
+                    </span>
+                  </Card>
+                )
+              }}
+              legends={
+                chartData.length === 1 || !enableLegend
+                  ? []
+                  : [
+                      {
+                        anchor: 'right',
+                        direction: 'column',
+                        itemWidth: 80,
+                        itemHeight: 20,
+                        translateX: 90,
+                        symbolSize: 10,
+                        symbolShape: 'circle',
+                      },
+                    ]
+              }
+              enableArea={enableArea}
+              areaOpacity={0.05}
+              enableSlices={enableSlices && 'x'}
+              onClick={(point) => {
+                if (!enableClick) return
+                const newVal = new Date(point.data.x).getTime()
+                if (clickFunction) clickFunction(newVal)
+              }}
+            />
+          )}
+        </Card>
+      </div>
+    )
+    // eslint-disable-next-line
+  }, [data, selectedStartDate, selectedEndDate, theme])
 }
 
 export default LineChart
